+++ src/main/java/com/maxmind/minfraud/request/Payment.java
@@ -244,6 +244,7 @@ public final class Payment extends AbstractModel {
         VIRTUAL_CARD_SERVICES,
         VME,
         VPOS,
+        WINDCAVE,
         WIRECARD,
         WORLDPAY;